Updated completion script installation with 'sudo tee' for Forge, Cast, and Anvil commands (#1002)
Revised the installation process for completion scripts for Forge, Cast, and Anvil commands to ensure proper permissions and prevent `permission denied` errors. Instead of direct file redirection with `>`, I used 'sudo tee' to write the output to the respective completion script files in `/usr/local/share/zsh/site-functions/`. This change ensures that the completion scripts are installed with superuser privileges, allowing them to function correctly in the Zsh shell.